pyproject.toml
[project]
name = "agent-builder-test"
version = "0.1.0"
description = ""
authors = [
    {name = "윤주연",email = "jyyun@didim365.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "annotated-types (==0.7.0)",
    "anyio (==4.9.0)",
    "build (==1.3.0)",
    "cachecontrol (==0.14.3)",
    "certifi (==2025.8.3)",
    "charset-normalizer (==3.4.2)",
    "cleo (==2.1.0)",
    "click (==8.2.1)",
    "colorama (==0.4.6)",
    "crashtest (==0.4.1)",
    "distlib (==0.4.0)",
    "dulwich (==0.22.8)",
    "fastapi (==0.116.1)",
    "fastjsonschema (==2.21.1)",
    "filelock (==3.18.0)",
    "findpython (==0.6.3)",
    "h11 (==0.16.0)",
    "httpcore (==1.0.9)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "installer (==0.7.0)",
    "jaraco-classes (==3.4.0)",
    "jaraco-context (==6.0.1)",
    "jaraco-functools (==4.2.1)",
    "keyring (==25.6.0)",
    "more-itertools (==10.7.0)",
    "msgpack (==1.1.1)",
    "packaging (==25.0)",
    "pbs-installer (==2025.7.23)",
    "pkginfo (==1.12.1.2)",
    "platformdirs (==4.3.8)",
    "poetry (==2.1.3)",
    "poetry-core (==2.1.3)",
    "prometheus-fastapi-instrumentator (==7.1.0)",
    "prometheus-client (==0.22.1)",
    "pydantic (==2.11.7)",
    "pydantic-core (==2.33.2)",
    "pyproject-hooks (==1.2.0)",
    "pywin32-ctypes (==0.2.3)",
    "rapidfuzz (==3.13.0)",
    "requests (==2.32.4)",
    "requests-toolbelt (==1.0.0)",
    "shellingham (==1.5.4)",
    "sniffio (==1.3.1)",
    "starlette (==0.47.2)",
    "tomlkit (==0.13.3)",
    "trove-classifiers (==2025.5.9.12)",
    "typing-inspection (==0.4.1)",
    "typing-extensions (==4.14.1)",
    "urllib3 (==2.5.0)",
    "uvicorn (==0.35.0)",
    "virtualenv (==20.33.0)",
    "zstandard (==0.23.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "redis (>=6.3.0,<7.0.0)",
    "clickhouse-sqlalchemy (>=0.3.2,<0.4.0)",
    "clickhouse-driver (>=0.2.9,<0.3.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "google-adk (==1.9.0)",
    "mcp (>=1.12.3,<2.0.0)",
    "deprecated (>=1.2.18,<2.0.0)",
    "adk (>=0.0.5,<0.0.6)",
    "aiohttp (>=3.12.15,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
